Move constants to be more idiomatic.
