+++ constants/table-of-contents.js
@@ -5,7 +5,7 @@ import React from 'react'
 
 
 export const tableContents = [
-        ["$20", "Funds an hour of RA learning"],
+    ["$20", "Funds an hour of RA learning"],
 		["$50", "Introduces a RA to a new skill"],
 		["$100", "-"],
 		["$1000", "Publishes an issue of JoER"],
